refactor: изменение архитектуры проекта для remote source
Ранее в проекте был один источник данных - репозиторий, где
данные были замоканы. Теперь данные в репозиторий приходят
из двух источников: локального и удаленного.
