+++ app/src/main/java/com/pasha/ytodo/domain/entities/TaskPriority.kt
@@ -1,4 +1,4 @@
-package com.pasha.ytodo.domain.models
+package com.pasha.ytodo.domain.entities
 
 import androidx.annotation.StringRes
 import com.pasha.ytodo.R
