+++ app/src/main/java/com/pasha/ytodo/domain/repositories/TodoItemRepositoryProvider.kt
@@ -1,7 +1,5 @@
 package com.pasha.ytodo.domain.repositories
 
-import com.pasha.ytodo.domain.models.TodoItem
-
 interface TodoItemRepositoryProvider {
     val todoItemsRepository: TodoItemsRepository
 }
