+++ app/src/main/java/com/pasha/ytodo/presentation/TodoItemViewModel.kt
@@ -1,12 +1,7 @@
 package com.pasha.ytodo.presentation
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.ViewModelProvider.AndroidViewModelFactory.Companion.APPLICATION_KEY
-import androidx.lifecycle.viewmodel.CreationExtras
-import com.pasha.ytodo.domain.models.TodoItem
-import com.pasha.ytodo.domain.repositories.TodoItemRepositoryProvider
-import com.pasha.ytodo.domain.repositories.TodoItemsRepository
+import com.pasha.ytodo.domain.entities.TodoItem
 
 
 class TodoItemViewModel : ViewModel() {
