+++ app/src/main/java/com/pasha/ytodo/presentation/edit/compose/priority/PriorityDropdownMenu.kt
@@ -22,7 +22,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
 import com.pasha.ytodo.R
-import com.pasha.ytodo.domain.models.TaskPriority
+import com.pasha.ytodo.domain.entities.TaskPriority
 import com.pasha.ytodo.presentation.edit.compose.share.OptionPickerTextField
