+++ app/src/main/java/com/pasha/ytodo/presentation/edit/compose/priority/PriorityMenuPreview.kt
@@ -11,7 +11,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.pasha.ytodo.domain.models.TaskPriority
+import com.pasha.ytodo.domain.entities.TaskPriority
 import com.pasha.ytodo.presentation.edit.compose.theme.TodoTheme
