+++ app/src/main/java/com/pasha/ytodo/presentation/edit/compose/screen/ChangebleTaskPart.kt
@@ -1,6 +1,6 @@
 package com.pasha.ytodo.presentation.edit.compose.screen
 
-import com.pasha.ytodo.domain.models.TaskPriority
+import com.pasha.ytodo.domain.entities.TaskPriority
 import java.time.LocalDateTime
 
 data class ChangeableTaskPart(
