+++ app/src/main/java/com/pasha/ytodo/presentation/edit/compose/screen/EditScreenContent.kt
@@ -26,7 +26,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.unit.dp
 import com.pasha.ytodo.R
-import com.pasha.ytodo.domain.models.TaskPriority
+import com.pasha.ytodo.domain.entities.TaskPriority
 import com.pasha.ytodo.presentation.edit.compose.deadline.DeadlinePicker
 import com.pasha.ytodo.presentation.edit.compose.priority.PriorityDropdownMenu
 import java.time.LocalDateTime
