+++ app/src/main/java/com/pasha/ytodo/presentation/edit/views/EditTaskState.kt
@@ -1,6 +1,6 @@
 package com.pasha.ytodo.presentation.edit.views
 
-import com.pasha.ytodo.domain.models.TaskPriority
+import com.pasha.ytodo.domain.entities.TaskPriority
 import java.time.LocalDateTime
 
 data class EditTaskState(
